refactor(csv): move encoding/csv to own top-level
