+++ csv/_io.ts
@@ -3,7 +3,7 @@
 // Copyright 2011 The Go Authors. All rights reserved. BSD license.
 // https://github.com/golang/go/blob/master/LICENSE
 // Copyright 2018-2023 the Deno authors. All rights reserved. MIT license.
-import { assert } from "../../_util/asserts.ts";
+import { assert } from "../_util/asserts.ts";
 
 export interface ReadOptions {
   /** Character which separates values.
